+++ src/game/client/proxyplayer.cpp
@@ -410,7 +410,7 @@ bool CPlayerLogoProxy::Init( IMaterial *pMaterial, KeyValues *pKeyValues )
 void CPlayerLogoProxy::OnBind( void *pC_BaseEntity )
 {
 	// Decal's are bound with the player index as the passed in paramter
-	int playerindex = (int)pC_BaseEntity;
+	int playerindex = (int)(intp)pC_BaseEntity;
 
 	if ( playerindex <= 0 )
 		return;
