+++ src/game/client/replay/vgui/replayperformanceeditor.cpp
@@ -1389,7 +1389,7 @@ static void Replay_GotoTick( bool bConfirmed, void *pContext )
 {
 	if ( bConfirmed )
 	{
-		int nGotoTick = (int)pContext;
+		int nGotoTick = (int)(intp)pContext;
 		CFmtStr fmtCmd( "demo_gototick %i\ndemo_pause\n", nGotoTick );
 		engine->ClientCmd_Unrestricted( fmtCmd.Access() );
 	}
