+++ src/game/server/info_camera_link.cpp
@@ -150,8 +150,8 @@ void PointCameraSetupVisibility( CBaseEntity *pPlayer, int area, unsigned char *
 		pCameraEnt->SetActive( false );
 	}
 	
-	int nNext;
-	for ( int i = g_InfoCameraLinkList.Head(); i != g_InfoCameraLinkList.InvalidIndex(); i = nNext )
+	intp nNext;
+	for ( intp i = g_InfoCameraLinkList.Head(); i != g_InfoCameraLinkList.InvalidIndex(); i = nNext )
 	{
 		nNext = g_InfoCameraLinkList.Next( i );
