+++ src/public/tier1/checksum_sha1.h
@@ -28,6 +28,8 @@
 		34AA973C D4C4DAA4 F61EEB2B DBAD2731 6534016F
 */
 
+#include "tier0/platform.h"
+
 #if !defined(_MINIMUM_BUILD_)
 #include <stdio.h>  // Needed for file access
 #if defined( _PS3 )
@@ -44,7 +46,7 @@
 typedef union
 {
 	unsigned char c[64];
-	unsigned long l[16];
+	uint32 l[16];
 } SHA1_WORKSPACE_BLOCK;
 
 // SHA1 hash
@@ -77,8 +79,8 @@ public:
 	~Minimum_CSHA1() ;	// no virtual destructor's in the minimal builds !
 #endif	
 
-	unsigned long m_state[5];
-	unsigned long m_count[2];
+	uint32 m_state[5];
+	uint32 m_count[2];
 	unsigned char m_buffer[64];
 	unsigned char m_digest[k_cubHash];
 
@@ -99,7 +101,7 @@ public:
 
 private:
 	// Private SHA-1 transformation
-	void Transform(unsigned long state[5], unsigned char buffer[64]);
+	void Transform(uint32 state[5], unsigned char buffer[64]);
 
 	// Member variables
 	unsigned char m_workspace[64];
