+++ src/public/vgui/Cursor.h
@@ -40,7 +40,7 @@ enum CursorCode
 	dc_alwaysvisible_pop,
 };
 
-typedef unsigned long HCursor;
+typedef uint32 HCursor;
 
 }
