+++ src/public/vgui/IInput.h
@@ -21,7 +21,7 @@ namespace vgui
 {
 
 class Cursor;
-typedef unsigned long HCursor;
+typedef uint32 HCursor;
 
 #define VGUI_GCS_COMPREADSTR                 0x0001
 #define VGUI_GCS_COMPREADATTR                0x0002
